MFC	r251212, r251480:
GCC: bring back experimental support for amdfam10/barcelona CPUs.

This is a revised subset of the support initially brought in in r236962
and later reverted. The collateral efects seem to have disappeared but
it is still recommended to set the CPUTYPE with caution.

While here also install AMD intrinsics header ammintrin.h

Reviewed by:	jkim (ages ago)