MFC: r261655
Refer newfs and growfs users to fsck_ffs instead of
fsck, the latter does not accept the referred to "-b" flag.

The part of the change pertaining newfs has already been merged
to stable/9 in r237505, so only record that part of the merge.

PR:		82720
Submitted by:	David D.W. Downey