Map the LAW for the RouterBoard's NAND LBC memory.
Without creating a LAW entry, any access to the NAND hangs the CPU.

The original intent was to add a quirk to map all of the RouterBoard's LBC
address space in one shot, which would fix it for both NAND and the CF, and
that's probably still in the cards.  However, for now, this makes NAND usable
without a separate hack.

Things left before the RouterBoard can run standalone:
* Add partitions to the NAND (not specified by the FDT, which we currently
  require).
* Create a YAFFS partition for the kernel.  The Mikrotik boot loader requires a
  4MB partition at the beginning of NAND, with a file called 'kernel' in the
  root.