Add (commented) 'vital' flag to the runtime packages.  Support for it is expected in pkg-1.8.0, which will cause 'pkg del -afy' to not destroy a system by forcefully removing everything.
As there are valid use cases for doing such (test jails, for example),
it will be overrideable.

In addition, ensure /dev and /tmp exist, as /dev is not created by
default currently, and pkg(8) requires /tmp to exist, which is also
not created by default.

This needs to be moved to the clibs package, since runtime depends on
that package, however I need to first verify that it will work as it
is expected.

Sponsored by:	The FreeBSD Foundation