MFC r302522, r302591-r302592 (by jmcneill)
r302522:
Align descriptors and data buffers to 32 bits. This restriction is
described in the A20 (and later) user manuals.

r302591:
Add support for arm64. The allwinner_soc_family() function is not available
on arm64 and all SoCs using the old FIFO register location are 32-bit only,
so unconditionally use the new location for arm64.

Reviewed by:	andrew, manu

r302592:
Return early from bus_dmamap_load callback if the error indicator is set.

Reviewed by:	andrew, manu